requirements.txt
django==2.2.1
djangorestframework==3.12.2